requirements.txt
fastapi
uvicorn
python-multipart
watchfiles
moviepy==1.0.3
opencv-python
numpy
Pillow
tqdm
python-dotenv